File Submission Function Added
Add the createSubmission method to backend to create a new submission
in the database. Add the request that sends the link, date, team
information and file information to the backend. Check in the front end
if the file is selected. If not, raise an error. Add the routers to
the corresponding route file.