Register supporter
The TFF admin now has the ability to send keys to supporters. These
keys allow supporters to sign up and access certain features
of the app.

This functionality is an important addition to the platform,
as it allows the admin to easily and effectively manage the
supporter base and grant access to system as needed.