Get support request status
This page allows supporters to view the status of their support
requests. If a request has been accepted by the team admin,
it will be marked as accepted on the page. If a request has
been rejected, it will be marked as rejected.

This feature is designed to help supporters stay informed about the
status of their support requests and ensure that they are aware of
any updates or changes.