Show sponsor budget
This feature allows users to view the sponsor budget for each
team on the “teams page”. If a team has a sponsor, the
amount of money that the sponsor has given to the team
will be displayed on this page. This allows all users who
visit the site to easily see the sponsor budgets for each team.