Review support requests
On the “review support request page”, we have implemented front-end
functionality that enables team admins to review incoming support
requests and decide whether to approve or reject them. This feature
allows team admins to view the amount of financial support received
and make informed decisions about whether to accept or reject
the request.

The goal of this feature is to provide organization admins(supporters)
with a clear and formal process for requesting support, and to ensure
that team admins have the information they need to make informed
decisions about these requests.